Make the betting round general, implement second betting round logic
Before, the betting round function was specific to pre flop betting.
Update it to support all betting rounds, and have a second setup
function for the later betting rounds.

Update the allIn logic to count a player as no longer in the hand (they
cannot perform another action so this makes sense).