Finish rest of betting, add a set for player names to avoid confusion
The betting logic is pretty straightforward, am able to reuse the
betting logic from the second round.

Add a set of player names to the common variables and add a check when
setting player names that the name is not already in use.

Clean up the cleanUpBettingRound function a bit.